Clean up file structure a bit based on review
- Tombstone stuff is now in tombstone.rs
- record_types.rs just includes type definitions for known record formats.
- service code is now in service.rs
- lib.rs just has crate decls, module decls, and exports.
